pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "brainvault"
version = "0.1.0"
description = "The Persistent State Infrastructure for the AI Era – pluggable DB + Markdown hybrid knowledge OS"
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.10"
keywords = ["ai", "knowledge-base", "mcp", "llm", "sqlite", "postgresql", "s3"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Database",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "click>=8.1",
    "PyYAML>=6.0",
    "rich>=13.0",
]

[project.optional-dependencies]
s3 = ["boto3>=1.28"]
postgresql = ["psycopg2-binary>=2.9"]
all = ["boto3>=1.28", "psycopg2-binary>=2.9"]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "pytest-mock>=3.12",
    "moto[s3]>=5.0",
    "ruff>=0.4",
    "mypy>=1.10",
    "types-PyYAML>=6.0",
]

[project.scripts]
brainvault = "brainvault.cli.main:cli"

[project.urls]
Homepage = "https://github.com/tanganke/BrainVault"
Issues = "https://github.com/tanganke/BrainVault/issues"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
brainvault = ["templates/**/*", "templates/**/.*", "templates/**/.gitignore"]

[tool.ruff]
src = ["src"]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.10"
strict = false
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"
